+++ cuda_modules/matrix_mult_trans.cu
@@ -14,7 +14,7 @@ extern "C" {
 #endif
 
 // CUDA Kernel
-__global__ void matrixMulTrans(float* C, float* A, float* B, int wA, int resW, int resH, int resultWidth, int resultSize)
+__global__ void matrixMulTrans(double* C, double* A, double* B, int wA, int resW, int resH, int resultWidth, int resultSize)
 {
 	int x = threadIdx.x + (blockIdx.x * resW);
 	int y = threadIdx.y + (blockIdx.y * resH);
@@ -24,7 +24,7 @@ __global__ void matrixMulTrans(float* C, float* A, float* B, int wA, int resW, i
 	if (resultPos < resultSize && x < resultWidth) {
 		// value stores the element that is 
 		// computed by the thread
-		float value = 0;
+		double value = 0;
 		for (int i = 0; i < wA; ++i)
 		{
 			value += A[y * wA + i] * B[x * wA + i];
