+++ cuda_modules/matrix_pow_two.cu
@@ -14,7 +14,7 @@ extern "C" {
 #endif
 
 // CUDA Kernel
-__global__ void matrixPowTwo(float* A, int resW, int resH, int width, int finalSize)
+__global__ void matrixPowTwo(double* A, int resW, int resH, int width, int finalSize)
 {
 	int x = threadIdx.x + (blockIdx.x * resW);
         int y = threadIdx.y + (blockIdx.y * resH);
