+++ cuda_modules/matrix_sigmoid_gradient.cu
@@ -14,7 +14,7 @@
 extern "C" {
 #endif
 
-__global__ void matrixSigmoidGrad(float* A, int resW, int resH, int width, int finalSize)
+__global__ void matrixSigmoidGrad(double* A, int resW, int resH, int width, int finalSize)
 {
 	int x = threadIdx.x + (blockIdx.x * resW);
         int y = threadIdx.y + (blockIdx.y * resH);
@@ -22,7 +22,7 @@ __global__ void matrixSigmoidGrad(float* A, int resW, int resH, int width, int f
 
 	if (resultPos < finalSize && x < width) {
 		//printf("IN Block %d - %d, wA: %d thread %d - %d Val: %f resultPos: %d finalSize: %d\n", x, y, wA, threadIdx.x, threadIdx.y, A[resultPos], resultPos, finalSize);
-		float s = 1 / (1 + pow(M_E, (double)(-1 * A[resultPos])));
+		double s = 1 / (1 + pow(M_E, (double)(-1 * A[resultPos])));
 		A[resultPos] = s * (1 - s);
 	}
 }
